Use 2to3 to port code to Python 3.
I tested the creation of formulas again, everything seems to be in
order. The change is just about migrating off of Python 2's print syntax
and semantics and onto Python 3. This makes it easier to use the script
when on more recent Linux distributions where Python 2 is slowly
deprecated more and more and Python 3 has become the default.

The command used for this port was `2to3 -w qbffam.py`. The package for
this is 2to3 in ubuntu.